Change it all to be a unittest.
