+++ koku/masu/test/database/test_partitions.py
@@ -2,16 +2,34 @@
 # Copyright 2023 Red Hat Inc.
 # SPDX-License-Identifier: Apache-2.0
 #
-"""Test that expected tables are partitioned correctly."""
-from unittest.mock import Mock
-from unittest.mock import patch
+"""Test that django tables are partitioned correctly."""
+from django.apps import apps
+from django.db import connections
+from django.db import DEFAULT_DB_ALIAS
 
-from masu.management.commands.create_partition_check import Command
-from masu.management.commands.create_partition_check import get_django_partitioned_models
-from masu.management.commands.create_partition_check import is_model_partitioned
 from masu.test import MasuTestCase
-from reporting.provider.ocp.models import OCPPodSummaryByNodeP
-from reporting.provider.ocp.models import OpenshiftCostCategory
+
+
+def is_model_partitioned(instance):
+    """Queries postgres to see if the table is partitioned."""
+    connection = connections[DEFAULT_DB_ALIAS]
+    table_name = instance._meta.db_table
+    check_partition_query = """
+    SELECT nmsp_parent.nspname AS parent_schema
+    FROM pg_inherits
+    JOIN pg_class parent
+        ON pg_inherits.inhparent = parent.oid
+        JOIN pg_class child
+        ON pg_inherits.inhrelid = child.oid
+        JOIN pg_namespace nmsp_parent
+        ON nmsp_parent.oid = parent.relnamespace
+        JOIN pg_namespace nmsp_child
+        ON nmsp_child.oid = child.relnamespace
+        WHERE parent.relname= %s
+        """
+    with connection.cursor() as cursor:
+        cursor.execute(check_partition_query, [table_name])
+        return bool(cursor.fetchone())
 
 
 class TestPartitionCheck(MasuTestCase):
@@ -32,48 +50,21 @@ class TestPartitionCheck(MasuTestCase):
 
         Example:
         - https://github.com/project-koku/koku/pull/4737"""
-        partitioned_models = get_django_partitioned_models()
+        # Collect models from django
+        partitioned_models = []
+        for model in apps.get_models():
+            if hasattr(model, "PartitionInfo"):
+                partitioned_models.append(model)
+
         if not partitioned_models:
             return
 
         # Check if models are actually partitioned in PostgreSQL
         for model in partitioned_models:
             with self.subTest(model=model):
-                self.assertTrue(is_model_partitioned(model))
-
-
-class TestPartitionCheckCommand(MasuTestCase):
-    def setUp(self):
-        self.command = Command()
-        self.stdout = Mock()
-        self.command.stdout = self.stdout
-        self.partitioned_model = OCPPodSummaryByNodeP
-        self.unpartitioned_model = OpenshiftCostCategory
-
-    def test_get_django_partitioned_models(self):
-        """Test that a partitioned model is in the return."""
-        partitioned_models = get_django_partitioned_models()
-        self.assertIn(self.partitioned_model, partitioned_models)
-
-    def test_is_model_partitioned(self):
-        """Test that the is_model_parttioned returns correct boolean."""
-        is_partitioned = is_model_partitioned(self.partitioned_model)
-        self.assertTrue(is_partitioned)
-        is_partitioned = is_model_partitioned(self.unpartitioned_model)
-        self.assertFalse(is_partitioned)
-
-    @patch("masu.management.commands.create_partition_check.get_django_partitioned_models")
-    @patch("masu.management.commands.create_partition_check.is_model_partitioned")
-    def test_handle(self, mock_is_model_partitioned, mock_get_django_partitioned_models):
-        """Test the handler for expected success & error."""
-        mock_get_django_partitioned_models.return_value = []
-        mock_is_model_partitioned.return_value = True
-        self.command.handle()
-        self.stdout.write.assert_called_with(self.command.style.SUCCESS("No models with PartitionInfo found."))
-        mock_get_django_partitioned_models.return_value = [self.unpartitioned_model]
-        mock_is_model_partitioned.return_value = False
-        self.stdout.reset_mock()
-        self.command.handle()
-        self.stdout.write.assert_called_with(
-            self.command.style.ERROR(f"Model {self.unpartitioned_model.__name__} is not partitioned.")
-        )
+                msg = (
+                    "Model {model.__name__} is not partitioned.\n"
+                    "\nAre `set_pg_extended_mode` and `unset_pg_extended_mode`"
+                    " used in the model's migration?"
+                )
+                self.assertTrue(is_model_partitioned(model), msg)
